requirements.txt
pyqudt>=1.1.0
